File reorg + API version check
