+++ spec/fixtures/bad_api_version.rb
@@ -1,0 +1,5 @@
+api_version 0
+
+filter do |event|
+  [event]
+end
