+++ spec/fixtures/no_api_version.rb
@@ -1,0 +1,3 @@
+filter do |event|
+  [event]
+end
